pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tcs-smart-analyzer"
version = "1.2.0"
description = "Offline desktop analyzer for TCS calibration data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "jinja2>=3.1",
    "scipy>=1.14",
    "pypinyin>=0.54",
    "python-docx>=1.1",
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.8",
]
mdf = [
    "asammdf>=8.0",
    "cantools>=39.4",
    "python-can>=4.4",
]
full = [
    "PySide6>=6.8",
    "asammdf>=8.0",
    "cantools>=39.4",
    "python-can>=4.4",
]

[project.scripts]
tcs-smart-analyzer = "tcs_smart_analyzer.main:main"
tcs-smart-analyzer-cli = "tcs_smart_analyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
tcs_smart_analyzer = ["config/*.json"]

[tool.setuptools.packages.find]
where = ["src"]
